+++ crates/veryl/src/cmd_build.rs
@@ -1,3 +1,4 @@
+use crate::cmd_check::CheckError;
 use crate::utils;
 use crate::OptBuild;
 use miette::{IntoDiagnostic, Result, WrapErr};
@@ -5,6 +6,7 @@ use std::fs;
 use std::fs::OpenOptions;
 use std::io::Write;
 use std::time::Instant;
+use veryl_analyzer::Analyzer;
 use veryl_emitter::Emitter;
 use veryl_metadata::{Metadata, Target};
 use veryl_parser::Parser;
@@ -25,7 +27,10 @@ impl CmdBuild {
             self.opt.files.clone()
         };
 
+        let mut all_pass = true;
         let now = Instant::now();
+
+        let mut check_error = CheckError::default();
 
         for file in &files {
             self.print(&format!(
@@ -35,6 +40,17 @@ impl CmdBuild {
 
             let input = fs::read_to_string(file).into_diagnostic().wrap_err("")?;
             let parser = Parser::parse(&input, file)?;
+
+            let mut analyzer = Analyzer::new(&input);
+            let errors = analyzer.analyze(&parser.veryl);
+            if !errors.is_empty() {
+                all_pass = false;
+
+                for error in errors {
+                    check_error.related.push(error);
+                }
+            }
+
             let mut emitter = Emitter::new(metadata);
             emitter.emit(&parser.veryl);
 
@@ -64,7 +80,11 @@ impl CmdBuild {
             elapsed_time.as_millis()
         ));
 
-        Ok(true)
+        if check_error.related.is_empty() {
+            Ok(all_pass)
+        } else {
+            Err(check_error.into())
+        }
     }
 
     fn print(&self, msg: &str) {