+++ testcases/sv/09_struct_enum.sv
@@ -8,8 +8,14 @@ module Module09 ;
 
     // enum declaration
     typedef enum logic [2-1:0] {
-        X = 1,
-        Y = 2,
-        Z
+        B_X = 1,
+        B_Y = 2,
+        B_Z
     } B;
+
+    A a ;
+    B b ;
+
+    assign a.a = 1;
+    assign b   = B_X;
 endmodule