+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * [Changed] modport separator from `.` to `::` [#65](https://github.com/dalance/veryl/issues/65)
 * [Changed] `[package]` in Veryl.toml to `[project]` [#82](https://github.com/dalance/veryl/issues/82)
 * [Added] type cast [#48](https://github.com/dalance/veryl/issues/48)
+* [Changed] enum member scope [#66](https://github.com/dalance/veryl/issues/66)
 
 ## [v0.1.14](https://github.com/dalance/veryl/compare/v0.1.13...v0.1.14) - 2023-01-12